Fixed error in parsing replies to netlink requests. Some requests result in multiple replies, sent in separate messages. This change keeps parsing replies until a done or error message is received.
Before this change, subsequent netlink calls could erroneously receive
replies for previous requests. This resulted in spurious error messages
in the log. Current code tolerated those unexpected replies.

Additionally: turned tabs into spaces in the affected files, changed
error messages in uvpn-ip-config.cc which were wrong, and addressed
minor warning in ip-addresses.cc.

+++ src/ip-addresses.cc
@@ -263,7 +263,7 @@ bool IPRange::GetIndexAddress(IPAddress* address, int* index) const {
 }
 
 void IPRange::Swap(IPRange* range) {
-  uint32_t tmp_cidr(cidr_);
+  unsigned short int tmp_cidr(cidr_);
   uint32_t tmp_size(size_);
   IPAddress* tmp_address(address_);
